Unbound: collision is integral — s32 vertices, bounds, plane distances and water boxes
The extent work made collision f32 to reach a +/-1048576 world. s32 reaches it just as
well and means a scene's geometry is exactly what its author placed: an editor snapping
to whole units gets back what it wrote, with no seam where two surfaces that should meet
sit a fraction apart. Vanilla kept collision integral and actor positions f32; this
restores that split with the range lifted.

collision.bin gains v3 (unbound/collision/3, SPEC 4.4.1): v2 with the two floating-point
fields made integral, so the record sizes are unchanged. v1 and v2 still load; a v2
document is ROUNDED, not truncated, so a vertex written as 99.9999 becomes 100 rather
than 99. The converter emits v3.

CollisionPoly.dist is derived from a unit normal, so it is fractional even when every
vertex is whole. It is rounded, which displaces a plane by at most half a unit --
precisely what vanilla did storing it in an s16. Likewise the dyna bake, which
transforms a moving actor's collision into world space every frame: those were Vec3s in
vanilla and are Vec3i now, so a moving platform's collision quantises to whole units as
it always did. Math_Vec3f_ToVec3i rounds for exactly this reason; truncating would drag
every platform's collision toward the origin.

Two mistakes worth recording, both the same shape and neither caught by the compiler:

- CollisionPoly and WaterBox are declared twice, in z64bgcheck.h and in the C++
  resource mirror, each carrying a comment saying it must mirror the other. Changing only
  the C side left the exporter casting a float dist to uint32_t: positive values survived,
  negative ones became 0, and half of every floor in the game vanished.
- ~20 functions in z_bgcheck.c took Vec3f* vtxList and were handed the widened Vec3i*,
  reading coordinates at the wrong offsets so no floor was ever found.

-Wincompatible-pointer-types did not flag the second, despite being enabled and despite
warning on a minimal reproduction of the same assignment; e1230338b claims more for that
flag than has been demonstrated. What did find both was decoding the generated
collision.bin and checking that every polygon's vertices satisfy its own plane equation,
against a known-good file as a control. Residuals are now all under 0.7 units (the
expected dist rounding) where the broken build had 43% in the thousands.

+++ soh/include/z64bgcheck.h
@@ -64,7 +64,9 @@ typedef struct {
     Vec3s normal; // Unit normal vector
                   // Value ranges from -0x7FFF to 0x7FFF, representing -1.0 to 1.0; 0x8000 is invalid
 
-    f32 dist; // Plane distance from origin along the normal. // SOH [Unbound] s16 -> f32 (world extent)
+    s32 dist; // Plane distance from origin along the normal. // SOH [Unbound] s16 -> s32 (world extent).
+              // Derived from a unit normal, so it is fractional even when every vertex is integral; it is
+              // rounded, exactly as vanilla rounded it into an s16.
 } CollisionPoly;
 
 typedef struct {
@@ -73,14 +75,14 @@ typedef struct {
     /* 0x04 */ Vec3s* camPosData;
 } CamData;
 
-// SOH [Unbound] Widened from the N64 0x10-byte layout: extents are f32 (world extent) and the packed
+// SOH [Unbound] Widened from the N64 0x10-byte layout: extents are s32 (world extent) and the packed
 // `properties` word is unpacked into fields (WaterBox_UnpackProperties for legacy data).
 typedef struct {
-    f32 xMin;
-    f32 ySurface;
-    f32 zMin;
-    f32 xLength;
-    f32 zLength;
+    s32 xMin;
+    s32 ySurface;
+    s32 zMin;
+    s32 xLength;
+    s32 zLength;
     s32 camera;       // CamData index (was 8 bits)
     s32 lightSetting; // lighting settings index (was 5 bits)
     s32 room;         // -1 = all rooms (was 6 bits, 0x3F)
@@ -118,12 +120,12 @@ void WaterBox_UnpackProperties(WaterBox* waterBox, u32 properties);
 }
 #endif
 
-// SOH [Unbound] Widened from the N64 layout: bounds and vertices are f32 (world extent), counts are u32
+// SOH [Unbound] Widened from the N64 layout: bounds and vertices are s32 (world extent), counts are u32
 typedef struct {
-    Vec3f minBounds; // minimum coordinates of poly bounding box
-    Vec3f maxBounds; // maximum coordinates of poly bounding box
+    Vec3i minBounds; // minimum coordinates of poly bounding box
+    Vec3i maxBounds; // maximum coordinates of poly bounding box
     u32 numVertices;
-    Vec3f* vtxList;
+    Vec3i* vtxList;
     u32 numPolygons;
     CollisionPoly* polyList;
     SurfaceType* surfaceTypeList;
@@ -193,7 +195,8 @@ typedef struct {
     void* retiredBuffers[DYNA_RETIRED_BUFFERS_MAX];
     s32 retiredCount;
     CollisionPoly* polyList;
-    Vec3f* vtxList; // s16 -> f32 (world extent)
+    Vec3i* vtxList; // s16 -> s32 (world extent). Baked world-space, so these quantise to whole units
+                    // each frame as they did in vanilla.
     DynaSSNodeList polyNodes;
     s32 polyNodesMax;
     s32 polyListMax;
